// src/app/interfaces/quest-encounter.ts
export interface QuestEncounter {
    monsterName: string;
    hitpoints: number;
    attack: number;
    defense: number;
    MonsterAttack1: string;
    MonsterAttack2: string;
    dodgeAttack: string;
    MonsterFlee: string;
    uniqueAction?: string;
    questCombat1: string;
    questCombat2: string;
    questCombat3: string;
    questdialogue1: string;
    questdialogue2: string;
    questdialogue3: string;
    questButton1: string;
    questButton2: string;
    questName: string;
}